+++ src/headings.ts
@@ -10,6 +10,7 @@ import {
     MAX_HEADER_DEPTH,
     getCurrentScope,
     getCursorPosition,
+    getCursorPositions,
     moveCursorToEndOfLine,
     getLine,
     getStartOfCurrentScope,
@@ -103,4 +104,59 @@ export function toggleLineAndHeading(textEditor: TextEditor, edit: TextEditorEdi
         addNewlineOnNonemptyLine: false,
         goToEndOfLine: false,
     });
+}
+
+export function demoteHeading(textEditor: TextEditor, edit: TextEditorEdit) {
+    const positions = getCursorPositions();
+    if (positions === null) {
+        return;
+    }
+
+    const document = textEditor.document;
+    let demotedSomeLine: boolean = false;
+    let position: Position;
+    let startOfHeading: string | null;
+    for (let i = 0; i < positions.length; i++) {
+        position = positions[i];
+        startOfHeading = getStartOfHeading(getLine(document, position));
+        if (startOfHeading === null || startOfHeading.length >= MAX_HEADER_DEPTH) {
+            continue;
+        }
+
+        demotedSomeLine = true;
+        edit.insert(position.with({ character: 0 }), HEADER_SYMBOL);
+    }
+
+    if (!demotedSomeLine) {
+        window.showWarningMessage("No line to demote");
+    }
+}
+
+export function promoteHeading(textEditor: TextEditor, edit: TextEditorEdit) {
+    const positions = getCursorPositions();
+    if (positions === null) {
+        return;
+    }
+
+    const document = textEditor.document;
+    let promotedSomeLine: boolean = false;
+    let position: Position;
+    let startOfHeading: string | null;
+    for (let i = 0; i < positions.length; i++) {
+        position = positions[i];
+        startOfHeading = getStartOfHeading(getLine(document, position));
+        if (startOfHeading === null || startOfHeading.length === 1) {
+            continue;
+        }
+
+        promotedSomeLine = true;
+        edit.delete(new Range(
+            position.with({ character: 0 }),
+            position.with({ character: 1 })
+        ));
+    }
+
+    if (!promotedSomeLine) {
+        window.showWarningMessage("No line to promote");
+    }
 }